fix: test util TempRepo, add posttitle set conf path
