+++ __test__/test_cases/isubo.js
@@ -33,3 +33,33 @@ export async function create_one_post(cb) {
 
   return lastRet;
 };
+
+export async function update_one_post(cb) {
+  const issue_number = 1;
+  const tempRepo = new TempRepo();
+  tempRepo.copy((preConf) => ({
+    ...preConf,
+    source_dir: tempRepo.tempSourceDir,
+    push_asset: enumPushAssetType.DISABLE
+  }));
+  const conf = tempRepo.conf;
+  // if(1) return console.info(conf)
+  const postParse = new PostParse({
+    path: path.join(tempRepo.tempSourceDir, 'license.md'),
+    conf
+  });
+  postParse.injectFrontmatter({
+    issue_number
+  });
+  const isubo = new Isubo({
+    conf,
+    cliParams: {
+      filename: 'license'
+    }
+  });
+  const ret = (await isubo.update()).pop();
+
+  cb && cb(ret);
+
+  return ret;
+}
